Phyton/requirements.txt
# Dependencias do SAMPLE API
delfinance-api-sdk
python-dotenv
